Routing issues
Changed it so that index.hbs will have no Ember Handlebar helpers
(link-to) and will instead hold a section that will  hold compiled
Ember templates (default to main.hbs). Currently getting a
initialTransition error in ember_prod.js and thus it is not loading
main.hbs into the section.

Also moved all of router.js to app.js because it is so small.