fixed routes, added electron, debugged to work
“Templates” are simply loaded on index.html as it seems loading the
compiled templates.js file did not add the template properly. Routing
(link-to) simply loads that template into it’s section on index.html.

electron was importing to node_modules and can be run using “electron
.” in the build directory. You will have to globally install
electron-prebuilt before this.

Would work in browser with jquery script on the bottom but with
electron I had to copy it into <head>.

Ember (A very simple use of it) now seems to be working with electron
properly.